Add files via upload
This is a updated version of the Distributed Hydrology Soil Vegetation Model with the Reservoir module (DHSVM_Res, Zhao et al., 2016, Advance in Water Resources). In this version, the surface routing scheme was updated to 8D flow direction, with a additional flow direction input file which can be processed by ArcGIS. Also, the reservoir release scheme was updated to make it more appropriate in simulations of detention reservoirs. 
DHSVM was developed by Wigmosta, et al. (1994), which is currently maintained by Pacific Northwest National Laboratory (PNNL) and University of Washington.